+++ src/apswversion.h
@@ -1,1 +1,1 @@
-#define APSW_VERSION "3.49.0.0"
+#define APSW_VERSION "3.49.1.0"